+++ src/EventTypeDefs.ts
@@ -1,6 +1,6 @@
 import type { Control } from './controls/control.class';
 import type { Point } from './point.class';
-import type { FabricObject } from './shapes/fabricObject.class';
+import type { FabricObject } from './shapes/FabricObject';
 import type { Group } from './shapes/group.class';
 import type { TOriginX, TOriginY, TRadian } from './typedefs';
 import type { saveObjectTransform } from './util/misc/objectTransforms';
