+++ src/mixins/canvas_straightening.mixin.ts
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { fabric } from '../../HEADER';
-import type { FabricObject } from '../shapes/fabricObject.class';
+import type { FabricObject } from '../shapes/FabricObject';
 
 fabric.util.object.extend(
   fabric.StaticCanvas.prototype,
