+++ src/mixins/collection.mixin.ts
@@ -1,5 +1,5 @@
 import { fabric } from '../../HEADER';
-import type { FabricObject } from '../shapes/fabricObject.class';
+import type { FabricObject } from '../shapes/FabricObject';
 import type { Constructor } from '../typedefs';
 
 export function createCollectionMixin<TBase extends Constructor>(Base: TBase) {
