+++ src/shapes/FabricObject.ts
