+++ src/shapes/path.class.ts
@@ -15,7 +15,7 @@ import {
   parsePath,
   type TPathSegmentsInfo,
 } from '../util/path';
-import { FabricObject, fabricObjectDefaultValues } from './fabricObject.class';
+import { FabricObject, fabricObjectDefaultValues } from './FabricObject';
 
 export class Path extends FabricObject {
   /**
