+++ src/util/misc/planeChange.ts
@@ -1,6 +1,6 @@
 import { iMatrix } from '../../constants';
 import type { Point } from '../../point.class';
-import type { FabricObject } from '../../shapes/fabricObject.class';
+import type { FabricObject } from '../../shapes/FabricObject';
 import type { TMat2D } from '../../typedefs';
 import { StaticCanvas } from '../../__types__';
 import { invertTransform, multiplyTransformMatrices } from './matrix';
